Adds a mechanism to share resources between suites
* When running suites in the context of a build tool, it is currently
impossible to share resources (http clients, databases) between several
suites.

This commit offers a non-intrusive, source-compatible way of doing it.

1. Users can implement `weaver.framework.GlobalResourcesInit` as
custom objects alongside their test suites. This interface is detected
as a task by build tools and managed by our framework implementation.
The method users have to implement provides a "store" (a shared state)
that they can use to store arbitrary objects in a cats.effect.Resource
program. That cats.effect.Resource is used by the framework to allocate
and clean-up resources the users might want to share between tests.
2. What the user can store is expressed in terms of a `ResourceTag`
that needs to be passed implicitly, with a default implementation based
on ClassTags and therefore subject to type-erasure. We try and protect
against type-erasure by preventing the use of this tag for types of kind
* -> * or (* -> *) -> *.
3. Users can then declare a constructor to their suite
(as opposed to having them be a static object). The constructor must
take a single `weaver.framework.GlobalResources` parameter. This object
offers a read-only interface to the shared store.

Shared resources get cleaned up after all the suites have run.