Throw on failed global (#275)
Fixes https://github.com/disneystreaming/weaver-test/issues/273

This change passes a promise to the IO loop, that should be filled as soon as the global resources are initialised or fail. The build tool waits for up to 2 minutes for that promise to be completed, in order to receive the tasks. 

Also changes the default monix Scheduler on jvm