Wait for resources release before ending tests
Fixes #473

A java Semaphore is created initially with 0 permits.  An additional cleanup Resource
is composed around the resources required by the tests.  When the test Resources
complete, the clean Resource releases a permit on the Semaphore.

SBTTask waits for this Semaphore to aquire a permit before wrapping things up.

Kudos to @Baccata for help with the fix.